Move Alpha class in its own module
